+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+#### Fixed
+
+- [#1364](https://github.com/rails-sqlserver/activerecord-sqlserver-adapter/pull/1364) Verify connection before retrieving the database version.
+
 ## v7.2.7
 
 #### Changed