+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## v7.1.12
 
 #### Fixed